added a note about /ON grounding
the RUN jumper on the adapter card will bypass the front panel /ON signal and always send a RUN signal to the CPU card.  However on front panels rev L and beyond, you still need the jumper on the front panel card if you want to enable the data LEDs and 7 segment displays.